Calculate Binding Affinities

Calculated binding affinities (Kd) from experimental data that contains dose and response information. Could be titrated concentration and resulting fluorescence, or percentage shift in a EMSA or something similar.

bio_binding(data, conc, resp, group = NULL, slope = 1)

Arguments

data

data.frame or tibble containing dose response data.

conc

Column with the concentration information.

resp

Column containing the response information.

group

allowed to be free (set slope = NA).

slope

For all 1:1 binding interactions, the slope for the log - logistic curve should be 1. If the interaction has some cooperativity or it is not a 1:1 binding interaction, then this value will change and should be

Value

a tibble

Examples

Puromycin %>%
  bio_binding(conc, rate, group = state)
#> # A tibble: 2 × 4
#>   state     data              model  coefs           
#>   <fct>     <list>            <list> <list>          
#> 1 treated   <tibble [12 × 2]> <drc>  <tibble [3 × 5]>
#> 2 untreated <tibble [11 × 2]> <drc>  <tibble [3 × 5]>