Extract Model Parameters to Columns

Extract Model Parameters to Columns

bio_coefs(data)

Arguments

data

Tibble containing a coefs column resulting from a model fit.

Value

tibble containing grouped variables and extracted parameters.

Examples

# Fitting MM models to the enzymatic data, pull out the relevant information.
Puromycin %>%
  bio_enzyme_rate(conc, rate, group = state) %>%
  bio_coefs()
#> # A tibble: 4 × 6
#>   state     term  estimate std.error statistic  p.value
#>   <fct>     <chr>    <dbl>     <dbl>     <dbl>    <dbl>
#> 1 treated   Vmax  213.       7.16        29.7  4.37e-11
#> 2 treated   Km      0.0641   0.00871      7.36 2.42e- 5
#> 3 untreated Vmax  160.       6.81        23.5  2.14e- 9
#> 4 untreated Km      0.0477   0.00842      5.67 3.07e- 4